requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
